PBL-V/PBL5/requirements.txt
streamlit
pillow
PyPDF2
docx2txt
requests
nibabel
numpy
opencv-python
biopython
reportlab
google-generativeai
scikit-learn
sentence-transformers
pydicom
